requirements.txt

dj-database-url 
django~=1.11.9
django-tastypie 
opencv-python 
pillow 
scikit-learn 
scipy 
numpy 

